pyproject.toml
[project]
name = "manimato-backend"
version = "0.1.0"
description = "FastAPI backend for LangChain-based RAG system"
readme = "README.md"
requires-python = ">=3.11,<3.12"

dependencies = [
    "fastapi>=0.115.14,<1.0.0",
    "langchain==0.1.20",
    "langchain-community>=0.0.38,<1.0.0",
    "langchain-openai==0.1.7",
    "sqlalchemy==2.0.41",
    "uvicorn[standard]==0.35.0",
    "manim>=0.18.1",
    "manimpango==0.6.0",
    "pycairo==1.25.0",
    "google-generativeai>=0.5.0",
    "numpy>=1.24.0",
    "passlib>=1.7.4",
    "jwt>=1.4.0",
    "pydantic[email]>=2.11.7",
    "pydantic-settings>=2.10.1",
    "bcrypt==3.2.2",
    "pyjwt>=2.10.1",
    "aiofiles>=24.1.0",
    "cloudinary>=1.44.1",
]

[dependency-groups]
dev = [
    "taskipy>=1.14.1",
    "ruff>=0.12.2",
    "pytest==8.2.2",
    "alembic>=1.16.2",
    "psycopg2-binary>=2.9,<3.0"
]
# dependency required only in production
prod = [
    "psycopg2>=2.9,<3.0",
]

[tool.taskipy.tasks]
dev = "uvicorn app.main:app --log-level debug"
start = "uvicorn app.main:app --host 0.0.0.0 --port 8000"
format = "ruff format ."
lint = "ruff check ."
lint-fix = "ruff check . --fix"
test = "pytest"
export = "uv export --format=requirements.txt --no-hashes > requirements.txt"
migrate = "alembic upgrade head"
clean-pycache = "echo 'Running: clean-pycache' && find ./app -type d -name '__pycache__' -print -exec rm -r {} +"

[tool.ruff]
line-length = 88
target-version = "py311"
exclude = [
    ".git",
    ".venv",
    "build",
    "dist",
    "migrations",
    "scripts/experimental.py",
    "legacy_code/"
]

[tool.ruff.lint]
extend-ignore = ["E203"]
extend-select = ["I"]
preview = true
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
exclude = [
    ".git",
    ".venv",
    "build",
    "dist",
    "migrations",
    "scripts/experimental.py",
    "legacy_code/"
]
